requirements.txt
fastAPI
uvicorn
psycopg2-binary
dotenv